movig components to entities
