+++ src/components/entities/Cakes/AllCakes.jsx
@@ -1,8 +1,8 @@
 import { Row, Card, Container, Col, Button } from "react-bootstrap";
-import { useCakeStore } from "../store/cakeStore";
+import { useCakeStore } from "../../store/cakeStore";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import DeleteConfirmation from "../shared/DeleteConfirmation";
+import DeleteConfirmation from "../../shared/DeleteConfirmation";
 
 const AllCakes = () => {
   const allCakes = useCakeStore((state) => state.cakeData);
